Targetting System, Player Data, Event handler, GUI
Implemented a targeting system to allow players to not hit specific entities (friends, pets, villagers, etc.), which can be customizable in the future. Started construction of player data to do this, which will evolve to a more robust system. 

Reworked event handlers for casting to address some bugs, and put safeguards in place to prevent unwanted casting. Massive issues initially, due to the server-side being unable to detect players opening their inventories, compared to other inventory/GUI events like chests.

Started working on a basic GUI for player data, not much there as of right now.